refactor(feature-flags): derive Wallet40Params from the Zod registry
Replace the hand-written 16-field `Wallet40Params` interface with
`Partial<Features["lwdWallet40"]["params"] & Features["lwmWallet40"]["params"]>`
so it stays in sync with the schema declarations in
@shared/feature-flags automatically.
