+++ features/platform/feature-flags/src/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./FeatureToggle";
