+++ features/platform/feature-flags/src/index.ts
@@ -1,5 +1,2 @@
-export { useFeature } from "./useFeature";
-export { useFeatureFlags } from "./useFeatureFlags";
-export { useHasLocallyOverriddenFeatureFlags } from "./useHasLocallyOverriddenFeatureFlags";
-export { FeatureToggle } from "./FeatureToggle";
-export * from "./walletFeaturesConfig";
+export * from "./components";
+export * from "./hooks";
