--- features/platform/feature-flags/src/useFeatureFlags.ts
@@ -1,7 +1,0 @@
-import { useSelector } from "react-redux";
-import type { FeatureFlagsState, Features } from "@shared/feature-flags";
-
-type WithFeatureFlags = { featureFlags: FeatureFlagsState };
-
-export const useFeatureFlags = (): Features =>
-  useSelector((state: WithFeatureFlags) => state.featureFlags.resolved);
