--- features/platform/feature-flags/src/walletFeaturesConfig/index.ts
@@ -1,3 +1,0 @@
-export { useWalletFeaturesConfig } from "./useWalletFeaturesConfig";
-export type { WalletPlatform, Wallet40Params, WalletFeaturesConfig } from "./types";
-export { FEATURE_FLAG_KEYS } from "./types";
